requirements.txt
pydantic>=2.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
textblob>=0.17.0
nltk>=3.7
colorama>=0.4.0
click>=8.0.0
PyYAML>=6.0.0
pathlib>=1.0.0
typing-extensions>=4.0.0
beautifulsoup4>=4.12.0